backend/requirements.txt
# Core Framework
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.13
langchain-openai>=0.0.5

# LLM Providers
openai>=1.10.0

# Vector Databases
pinecone>=5.0.0  # Updated: pinecone-client renamed to pinecone
chromadb>=0.4.22  # For local development

# Document Processing
pypdf>=4.0.0
pymupdf>=1.23.8
python-docx>=1.1.0
arxiv>=2.1.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Caching & Queue
redis>=5.0.1
celery>=5.3.6

# Monitoring & Logging
langsmith>=0.0.87
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
httpx>=0.26.0
aiohttp>=3.9.1

# Data Processing
numpy>=1.26.3
pandas>=2.1.4

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=22.0.0

# Code Quality
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.2
pre-commit>=3.6.0

# Security
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3

# Development
ipython>=8.20.0
jupyter>=1.0.0

